crates/doc: refactor inference into sub-module `shape`
doc::shape has a variety of other submodules, some of them public and
others not, which focus on particular features and capabilities of Shape.

No behavior changes, and no significant changes to tests,
though I moved quite a few things around.
